requirements.txt
Flask==1.0.2
PyJWT==1.7.1
pymongo==3.7.2
